+++ amethyst_controls/src/systems.rs
@@ -1,5 +1,8 @@
 use derive_new::new;
-use winit::{DeviceEvent, Event, Window, WindowEvent};
+use winit::{
+    event::{DeviceEvent, Event, WindowEvent},
+    window::Window,
+};
 
 #[cfg(feature = "profiler")]
 use thread_profiler::profile_scope;
@@ -132,7 +135,9 @@ impl<'a, 'b> SystemDesc<'a, 'b, FreeRotationSystem> for FreeRotationSystemDesc {
     fn build(self, world: &mut World) -> FreeRotationSystem {
         <FreeRotationSystem as System<'_>>::SystemData::setup(world);
 
-        let event_reader = world.fetch_mut::<EventChannel<Event>>().register_reader();
+        let event_reader = world
+            .fetch_mut::<EventChannel<Event<()>>>()
+            .register_reader();
 
         FreeRotationSystem::new(self.sensitivity_x, self.sensitivity_y, event_reader)
     }
@@ -149,12 +154,12 @@ impl<'a, 'b> SystemDesc<'a, 'b, FreeRotationSystem> for FreeRotationSystemDesc {
 pub struct FreeRotationSystem {
     sensitivity_x: f32,
     sensitivity_y: f32,
-    event_reader: ReaderId<Event>,
+    event_reader: ReaderId<Event<()>>,
 }
 
 impl<'a> System<'a> for FreeRotationSystem {
     type SystemData = (
-        Read<'a, EventChannel<Event>>,
+        Read<'a, EventChannel<Event<()>>>,
         WriteStorage<'a, Transform>,
         ReadStorage<'a, FlyControlTag>,
         Read<'a, WindowFocus>,
@@ -193,7 +198,9 @@ impl<'a, 'b> SystemDesc<'a, 'b, MouseFocusUpdateSystem> for MouseFocusUpdateSyst
     fn build(self, world: &mut World) -> MouseFocusUpdateSystem {
         <MouseFocusUpdateSystem as System<'_>>::SystemData::setup(world);
 
-        let event_reader = world.fetch_mut::<EventChannel<Event>>().register_reader();
+        let event_reader = world
+            .fetch_mut::<EventChannel<Event<()>>>()
+            .register_reader();
 
         MouseFocusUpdateSystem::new(event_reader)
     }
@@ -202,11 +209,11 @@ impl<'a, 'b> SystemDesc<'a, 'b, MouseFocusUpdateSystem> for MouseFocusUpdateSyst
 /// A system which reads Events and saves if a window has lost focus in a WindowFocus resource
 #[derive(Debug, new)]
 pub struct MouseFocusUpdateSystem {
-    event_reader: ReaderId<Event>,
+    event_reader: ReaderId<Event<()>>,
 }
 
 impl<'a> System<'a> for MouseFocusUpdateSystem {
-    type SystemData = (Read<'a, EventChannel<Event>>, Write<'a, WindowFocus>);
+    type SystemData = (Read<'a, EventChannel<Event<()>>>, Write<'a, WindowFocus>);
 
     fn run(&mut self, (events, mut focus): Self::SystemData) {
         #[cfg(feature = "profiler")]
@@ -255,16 +262,16 @@ impl<'a> System<'a> for CursorHideSystem {
 
         let should_be_hidden = focus.is_focused && hide.hide;
         if !self.is_hidden && should_be_hidden {
-            if let Err(err) = win.grab_cursor(true) {
+            if let Err(err) = win.set_cursor_grab(true) {
                 log::error!("Unable to grab the cursor. Error: {:?}", err);
             }
-            win.hide_cursor(true);
+            win.set_cursor_visible(false);
             self.is_hidden = true;
         } else if self.is_hidden && !should_be_hidden {
-            if let Err(err) = win.grab_cursor(false) {
+            if let Err(err) = win.set_cursor_grab(false) {
                 log::error!("Unable to release the cursor. Error: {:?}", err);
             }
-            win.hide_cursor(false);
+            win.set_cursor_visible(true);
             self.is_hidden = false;
         }
     }