+++ amethyst_window/src/monitor.rs
@@ -1,30 +1,30 @@
 use serde::{Deserialize, Serialize};
-use winit::{AvailableMonitorsIter, EventsLoop, MonitorId, Window};
+use winit::{event_loop::EventLoop, monitor::MonitorHandle, window::Window};
 
 /// A struct that can resolve monitors.
 /// Usually either a Window or an EventsLoop.
 pub trait MonitorsAccess {
     /// Returns an iterator over the available monitors
-    fn iter(&self) -> AvailableMonitorsIter;
-    /// Returns the `MonitorId` of the primary display
-    fn primary(&self) -> MonitorId;
+    fn iter(&self) -> Box<dyn Iterator<Item = MonitorHandle>>;
+    /// Returns the `MonitorHandle` of the primary display
+    fn primary(&self) -> MonitorHandle;
 }
 
-impl MonitorsAccess for EventsLoop {
-    fn iter(&self) -> AvailableMonitorsIter {
-        self.get_available_monitors()
+impl MonitorsAccess for EventLoop<()> {
+    fn iter(&self) -> Box<dyn Iterator<Item = MonitorHandle>> {
+        Box::new(self.available_monitors())
     }
-    fn primary(&self) -> MonitorId {
-        self.get_primary_monitor()
+    fn primary(&self) -> MonitorHandle {
+        self.primary_monitor()
     }
 }
 
 impl MonitorsAccess for Window {
-    fn iter(&self) -> AvailableMonitorsIter {
-        self.get_available_monitors()
+    fn iter(&self) -> Box<dyn Iterator<Item = MonitorHandle>> {
+        Box::new(self.available_monitors())
     }
-    fn primary(&self) -> MonitorId {
-        self.get_primary_monitor()
+    fn primary(&self) -> MonitorHandle {
+        self.primary_monitor()
     }
 }
 
@@ -40,16 +40,19 @@ pub struct MonitorIdent(u16, String);
 impl MonitorIdent {
     /// Get the identifier for current primary monitor.
     pub fn from_primary(monitors: &impl MonitorsAccess) -> Self {
-        Self::from_monitor_id(monitors, monitors.primary())
+        Self::from_monitor_handle(monitors, monitors.primary())
             .expect("Primary monitor not found in the list of all monitors")
     }
 
     /// Get the identifier for specific monitor id.
-    pub fn from_monitor_id(monitors: &impl MonitorsAccess, monitor_id: MonitorId) -> Option<Self> {
+    pub fn from_monitor_handle(
+        monitors: &impl MonitorsAccess,
+        monitor_handle: MonitorHandle,
+    ) -> Option<Self> {
         #[cfg(target_os = "ios")]
-        use winit::ios::windows::MonitorIdExt;
+        use winit::platform::ios::MonitorHandleExtIOS;
         #[cfg(target_os = "macos")]
-        use winit::os::macos::MonitorIdExt;
+        use winit::platform::macos::MonitorHandleExtMacOS;
         #[cfg(any(
             target_os = "linux",
             target_os = "dragonfly",
@@ -57,24 +60,24 @@ impl MonitorIdent {
             target_os = "netbsd",
             target_os = "openbsd"
         ))]
-        use winit::os::unix::MonitorIdExt;
+        use winit::platform::unix::MonitorHandleExtUnix;
         #[cfg(target_os = "windows")]
-        use winit::os::windows::MonitorIdExt;
+        use winit::platform::windows::MonitorHandleExtWindows;
 
-        let native_id = monitor_id.native_id();
+        let native_id = monitor_handle.native_id();
         monitors
             .iter()
             .enumerate()
             .find(|(_, m)| m.native_id() == native_id)
-            .and_then(|(i, m)| m.get_name().map(|name| Self(i as u16, name)))
+            .and_then(|(i, m)| m.name().map(|name| Self(i as u16, name)))
     }
 
     /// Select a monitor that matches this identifier most closely.
-    pub fn monitor_id(&self, monitors: &impl MonitorsAccess) -> MonitorId {
+    pub fn monitor_handle(&self, monitors: &impl MonitorsAccess) -> MonitorHandle {
         monitors
             .iter()
             .enumerate()
-            .filter(|(_, m)| m.get_name().map(|n| n == self.1).unwrap_or(false))
+            .filter(|(_, m)| m.name().map(|n| n == self.1).unwrap_or(false))
             .max_by_key(|(i, _)| (*i as i32 - i32::from(self.0)).abs() as u16)
             .map(|(_, m)| m)
             .unwrap_or_else(|| monitors.primary())